Added tests for comparison operators, refactored min/max/round built-in function tests
